Allow having a different interpreter for each path.
This should allow having custom interpreters based on the
package.yaml in Robocorp Code (this still needs to be done).

The idea is that the RobotFramework Language Server provides
a way to customize plugins through the robot.addPluginsDir
command, then, a client can provide an implementation for
the EPResolveInterpreter extension point, which allows
customizing an interpreter for a given path.