Make launches on the terminal by default (robocorp code and robot framework language server).
In general, the terminal is more user-friendly to users (Ctrl+C should work)
and it should allow us to handle hyperlinks in the future.

Note that on Robocorp Code launches a wrapper script is needed for doing the rcc launch to
discover the pid of the launches process (which is needed for the debug adapter
when doing non-debug launches).

For the Robot Framework Language Server this isn't needed as we always launch through a
wrapper python script which provides the pid already.