Work in progress: Improved Robot Output
This is the initial commit for adding a new Robot Output.
Although it supports loading an output.xml file and showing
it, the idea is having a new format more streamlined to
stream the contents of the run as it's happening while also
allowing for rotating the output if it grows too big.