Browser picker stability / $/webInspectorState message / fixes / refactorings / pre-built env updated.
Improved the stability of the code that ensures the picker is always active.
Added $/webInspectorState message to notify about changes in the browser/picking state.
Updated tests to deal with the new locator structure.
Updated messages in the language server side to return ActionResult (the inspectorView.ts takes care of dealing with the ids back and for with the webview).
Updated the used pre-built environments.